fix: rename module to template metrics
