+++ marketplace/wpp_templates/metrics/views.py
@@ -3,22 +3,22 @@ from rest_framework.response import Response
 from rest_framework.views import APIView
 from rest_framework import status
 
-from marketplace.wpp_templates.insights.dto import TemplateMetricsDTO
-from marketplace.wpp_templates.insights.serializers import TemplateVersionDataSerializer
-from marketplace.wpp_templates.insights.usecases.template_metrics import (
+from marketplace.wpp_templates.metrics.dto import TemplateMetricsDTO
+from marketplace.wpp_templates.metrics.serializers import TemplateVersionDataSerializer
+from marketplace.wpp_templates.metrics.usecases.template_metrics import (
     TemplateMetricsUseCase,
 )
 
 
-class TemplateVersionDataView(APIView):
+class TemplateMetricsView(APIView):
     """
-    View responsible for retrieving template insight data
+    View responsible for retrieving template metrics data
     based on gallery versions and a specified time range.
     """
 
     def post(self, request: Request, app_uuid: str) -> Response:
         """
-        Handles POST requests for retrieving template insights.
+        Handles POST requests for retrieving template metrics.
 
         Path param:
             app_uuid (str): UUID of the application
@@ -31,7 +31,7 @@ class TemplateVersionDataView(APIView):
             }
 
         Returns:
-            Response with status 200 and insight data.
+            Response with status 200 and metrics data.
         """
         serializer = TemplateVersionDataSerializer(data=request.data)
         serializer.is_valid(raise_exception=True)
